fix: repair broken base migration so db:migrate works from scratch
Two separate bugs in the migration chain were blocking anyone from
setting up the database from an empty schema with plain `yarn db:migrate`:

1. Every foreign key column in 0000_numerous_cargill.sql was typed
   `int`, but every referenced primary key is `serial` (bigint
   unsigned) per schema.ts. MySQL rejects FK constraints between
   mismatched types, so migration 0000 failed partway through and
   never got recorded as applied. Fixed all 10 affected FK columns
   (accounts.user_id, classification_results.*, device_settings.
   device_id, device_user.*, sensor_readings.device_id, sessions.
   user_id, telegram_logs.user_id, users.role_id) to bigint unsigned,
   in both the raw SQL and the matching snapshot metadata.

2. 0001_sigma_earthquake_fields.sql was hand-added without registering
   it in meta/_journal.json, so drizzle's migrator never knew it
   existed and silently skipped it even after fixing #1.

Verified end-to-end on a throwaway empty database: `yarn db:migrate`
now creates all 13 tables with correct FK types and the SIGMA columns,
with both migrations properly recorded in __drizzle_migrations.

+++ backend/src/database/migrations/0000_numerous_cargill.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `accounts` (
 	`id` varchar(255) NOT NULL,
-	`user_id` int,
+	`user_id` bigint unsigned,
 	`account_id` varchar(255) NOT NULL,
 	`provider_id` varchar(255) NOT NULL,
 	`access_token` text,
@@ -14,8 +14,8 @@ CREATE TABLE `accounts` (
 --> statement-breakpoint
 CREATE TABLE `classification_results` (
 	`id` serial AUTO_INCREMENT NOT NULL,
-	`sensor_reading_id` int,
-	`device_id` int,
+	`sensor_reading_id` bigint unsigned,
+	`device_id` bigint unsigned,
 	`label` varchar(255) NOT NULL,
 	`confidence` float NOT NULL,
 	`created_at` timestamp DEFAULT (now()),
@@ -24,7 +24,7 @@ CREATE TABLE `classification_results` (
 --> statement-breakpoint
 CREATE TABLE `device_settings` (
 	`id` serial AUTO_INCREMENT NOT NULL,
-	`device_id` int,
+	`device_id` bigint unsigned,
 	`initial_distance` float NOT NULL,
 	`alert_threshold` float NOT NULL,
 	`danger_threshold` float NOT NULL,
@@ -35,8 +35,8 @@ CREATE TABLE `device_settings` (
 --> statement-breakpoint
 CREATE TABLE `device_user` (
 	`id` serial AUTO_INCREMENT NOT NULL,
-	`user_id` int,
-	`device_id` int,
+	`user_id` bigint unsigned,
+	`device_id` bigint unsigned,
 	`created_at` timestamp DEFAULT (now()),
 	CONSTRAINT `device_user_id` PRIMARY KEY(`id`)
 );
@@ -64,7 +64,7 @@ CREATE TABLE `roles` (
 --> statement-breakpoint
 CREATE TABLE `sensor_readings` (
 	`id` serial AUTO_INCREMENT NOT NULL,
-	`device_id` int,
+	`device_id` bigint unsigned,
 	`recorded_at` datetime NOT NULL,
 	`temperature` double,
 	`humidity` double,
@@ -91,7 +91,7 @@ CREATE TABLE `sensor_readings` (
 --> statement-breakpoint
 CREATE TABLE `sessions` (
 	`id` varchar(255) NOT NULL,
-	`user_id` int,
+	`user_id` bigint unsigned,
 	`token` text NOT NULL,
 	`expires_at` timestamp NOT NULL,
 	`ip_address` text,
@@ -114,7 +114,7 @@ CREATE TABLE `system_settings` (
 CREATE TABLE `telegram_logs` (
 	`id` serial AUTO_INCREMENT NOT NULL,
 	`chat_id` varchar(255) NOT NULL,
-	`user_id` int,
+	`user_id` bigint unsigned,
 	`message` text NOT NULL,
 	`type` enum('sent','received') DEFAULT 'sent',
 	`created_at` timestamp DEFAULT (now()),
@@ -131,7 +131,7 @@ CREATE TABLE `users` (
 	`image` text,
 	`created_at` timestamp DEFAULT (now()),
 	`updated_at` timestamp DEFAULT (now()) ON UPDATE CURRENT_TIMESTAMP,
-	`role_id` int,
+	`role_id` bigint unsigned,
 	`telegram_chat_id` varchar(255),
 	CONSTRAINT `users_id` PRIMARY KEY(`id`),
 	CONSTRAINT `users_email_unique` UNIQUE(`email`)
